// src/components/types.ts
export interface EffectSettings {
  intensity: number;
  colorShift: number;
  pixelation: number;
  noise: number;
  waveAmplitude: number;
  waveFrequency: number;
  rgbSplit: number;
  hueRotate: number;
  scanlines: number;
  vhsTracking: number;
  signalInterference: number;
  saturation: number;
  contrast: number;
}